requirements.txt
streamlit==1.40.0
requests==2.32.3
python-dotenv==1.0.1
plotly==5.24.1
snowflake-ml-python==1.6.3
cryptography==44.0.0
PyJWT==2.10.1
streamlit-option-menu==0.4.0
pandas==2.2.3
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
